refactor: moving logo code in common.tsx
